Only set highlight if not previously set by user
On every window focus, the highlight would get reset to the default if
it was not specified in the plugin's setup.

This will allow for setting the highlight of the window separator to a
custom value that is set later in the user's config. Such as when
setting up a colorscheme after the plugin has already been loaded.